add buttons for the filters
