+++ dotcom-rendering/src/web/components/AutomaticFilterButton.importable.tsx
@@ -1,0 +1,18 @@
+import { Button } from '@guardian/source-react-components';
+
+type Props = {
+	label: string;
+};
+const AutomaticFilterButton = ({ label }: Props) => {
+	return (
+		<Button
+			onClick={() => {
+				console.log('button clicked', label);
+			}}
+		>
+			{label}
+		</Button>
+	);
+};
+
+export default AutomaticFilterButton;
